# spring-boot/rest-api-mongodb/openapi.yaml
openapi: 3.0.0
info:
  title: Product API
  description: API for managing products
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
  /products:
    get:
      summary: Get all products
      description: Retrieve a list of all products, optionally filtered by category.
      parameters:
        - name: category
          in: query
          description: Filter products by category
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product
      description: Create a new product with the provided details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: The created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{id}:
    get:
      summary: Get a product by ID
      description: Retrieve a product by its unique ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    head:
      summary: Check if a product exists
      description: Check if a product with the given ID exists.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product exists
        '404':
          description: Product does not exist
    patch:
      summary: Partially update a product
      description: Update specific fields of a product.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: The updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      summary: Update a product
      description: Replace a product with the provided details.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: The updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      summary: Delete a product
      description: Delete a product by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted successfully
    options:
      summary: Get allowed operations for a product
      description: Retrieve the allowed HTTP methods for a product.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Allowed operations
          headers:
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the product
        name:
          type: string
          description: The name of the product
        category:
          type: string
          description: The category of the product
        description:
          type: string
          description: The description of the product
        attributes:
          type: object
          additionalProperties:
            type: string
          description: Additional attributes of the product
        price:
          type: number
          format: double
          description: The price of the product
      required:
        - name
        - price